requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
jinja2==3.1.2
pandas
bleach
transformers
Flask-Caching
Flask-Compress
sentence-transformers
faiss-cpu
numpy
sqlalchemy[asyncio]
gunicorn
scipy
python-dotenv
psycopg2-binary
Flask-Migrate==4.0.5